+++ jsonargparse/_type_checking.py
@@ -2,6 +2,7 @@ from typing import TYPE_CHECKING
 
 __all__ = [
     "_ArgumentGroup",
+    "ActionsContainer",
     "ArgumentParser",
     "ruyamlCommentedMap",
 ]
@@ -9,6 +10,6 @@ __all__ = [
 if TYPE_CHECKING:  # pragma: no cover
     from ruyaml.comments import CommentedMap as ruyamlCommentedMap
 
-    from ._core import ArgumentParser, _ArgumentGroup
+    from ._core import ActionsContainer, ArgumentParser, _ArgumentGroup
 else:
     globals().update({k: None for k in __all__})