+++ jsonargparse/_typehints.py
@@ -227,7 +227,6 @@ class ActionTypeHint(Action):
         """
         if typehint is not None:
             if not self.is_supported_typehint(typehint, full=True):
-                self.is_supported_typehint(typehint, full=True)
                 raise ValueError(f"Unsupported type hint {typehint}.")
             if get_typehint_origin(typehint) == Union:
                 subtype_supported = [
@@ -274,6 +273,8 @@ class ActionTypeHint(Action):
             default_type = type(default)
             if not is_subclass(default_type, UnknownDefault) and self.is_subclass_typehint(default_type):
                 raise ValueError("Subclass types require as default either a dict with class_path or a lazy instance.")
+        elif ActionTypeHint.is_return_subclass_typehint(self._typehint) and inspect.isclass(default):
+            default = {"class_path": get_import_path(default)}
         return default
 
     @staticmethod
@@ -429,21 +430,6 @@ class ActionTypeHint(Action):
                         ActionTypeHint.discard_init_args_on_class_path_change(subparser, prev_sub_cfg, sub_cfg)
                     keys = keys[: num + 1] + [k for k in keys[num + 1 :] if not k.startswith(key + ".")]
             num += 1
-
-    @staticmethod
-    def delete_init_args_required_none(cfg_from, cfg_to):
-        for key, val in cfg_from.items(branches=True):
-            if isinstance(val, Namespace) and val.get("class_path") and val.get("init_args"):
-                skip_keys = [
-                    k
-                    for k, v in val.init_args.__dict__.items()
-                    if v is None and cfg_to.get(f"{key}.init_args.{k}") is not None
-                ]
-                if skip_keys:
-                    parser = ActionTypeHint.get_class_parser(val.class_path)
-                    for skip_key in skip_keys:
-                        if skip_key in parser.required_args:
-                            del val.init_args[skip_key]
 
     @staticmethod
     def delete_not_required_args(cfg_from, cfg_to):
@@ -1021,6 +1007,7 @@ def adapt_typehints(
                         sub_add_kwargs,
                         skip_args=num_partial_args,
                         partial_classes=partial_classes,
+                        prev_val=prev_val,
                     )
             except (ImportError, AttributeError, ArgumentError) as ex:
                 raise_unexpected_value(f"Type {typehint} expects a function or a callable class: {ex}", val, ex)
@@ -1169,8 +1156,6 @@ def subclass_spec_as_namespace(val, prev_val=None):
         val = Namespace({root_key: val})
         if isinstance(prev_val, str):
             prev_val = Namespace(class_path=prev_val)
-        elif inspect.isclass(prev_val):
-            prev_val = Namespace(class_path=get_import_path(prev_val))
     if isinstance(val, dict):
         val = Namespace(val)
     if "init_args" in val and isinstance(val["init_args"], dict):